Finish mission 1
The final commit related to the development of the First mission, including it's ending actions set.

Main fixes address the effort required to access certain nodes as well as readability of some texts within the game.

Biggest addition of the commit is the "missionEnd" action set, linking it up with the next mission.